feat(v0.7.2): engine overhaul + glitch filter + RAG skills
Core Engine Overhaul:
- Engine enum: add Local(HTTP) and Skill(name) variants, remove Copy derive
- Engine::default_enabled() with AION_ORCH_DISABLED_ENGINES env var
- Engine::cli_available() for CLI path detection
- call_skill_engine() — dynamic skill-as-engine execution
- call_http_ai_fallback() — HTTP fallback for local engine
- All builtins use dynamic engine selection over hardcoded lists

Security Layer (crates/glitch-filter):
- GlitchFilter: 1195 lines, 25 categories, 19 unit tests
- Sanitizer: NULL/DEL/C1/zero-width/Bidi/BOM/ANSI/r-flood defense
- Integrated into aion-forge-cli entrypoint

New Skills:
- skills/mcp_call/ — MCP tool invocation skill
- skills/rag_ingest/ — RAG knowledge base ingestion
- skills/rag_query/ — RAG query and retrieval
- skills/rag_status/ — RAG status monitoring

Types Expansion (aion-types):
- research module with evaluation/optimize/refine/spectral
- route_types module for enhanced routing types
- CapabilityRegistry improvements
- agent_message and agent_template modules

Infrastructure:
- MCP tool registration and config enhancements
- Registry and routing config updates
- Docs: system prompts analysis, innoforge usability rating
- .gitignore: exclude .skill-router/rag/ runtime data

+++ aion-router/src/metrics.rs
@@ -51,7 +51,7 @@ pub fn record_token_usage(
         "capability" => capability.to_string(),
         "provider" => provider.to_string()
     )
-    .increment(usage.prompt_tokens as u64);
+    .increment(usage.prompt_tokens);
 
     metrics::counter!(
         "skill_ai_completion_tokens_total",
@@ -59,7 +59,7 @@ pub fn record_token_usage(
         "capability" => capability.to_string(),
         "provider" => provider.to_string()
     )
-    .increment(usage.completion_tokens as u64);
+    .increment(usage.completion_tokens);
 
     metrics::counter!(
         "skill_ai_total_tokens_total",
@@ -67,7 +67,7 @@ pub fn record_token_usage(
         "capability" => capability.to_string(),
         "provider" => provider.to_string()
     )
-    .increment(usage.total_tokens as u64);
+    .increment(usage.total_tokens);
 
     if usage.cached_tokens > 0 {
         metrics::counter!(
@@ -76,6 +76,6 @@ pub fn record_token_usage(
             "capability" => capability.to_string(),
             "provider" => provider.to_string()
         )
-        .increment(usage.cached_tokens as u64);
+        .increment(usage.cached_tokens);
     }
 }
